Split route tests into separate python files
